requirements.txt
aider-chat
diff-match-patch
intervaltree
rope
llama-index
httpx
httpcore
openai > 1.3.1

